Removed some apis from export in container loader package (#6528)
