+++ packages/loader/container-loader/src/index.ts
@@ -7,5 +7,4 @@ export * from "./audience";
 export * from "./container";
 export * from "./deltaManager";
 export * from "./loader";
-export * from "./utils";
 export * from "./retriableDocumentStorageService";
